Made a lot of changes :cupid:

+++ controller/file.js
@@ -8,9 +8,12 @@ exports.upload=function(q,s){
     var editor_id = q.query.editorid;
     var target_path;
     console.log(q.header.host);
-    if(q.files.file.size===0){
+    if(q.files.file && q.files.file.size===0){
         fs.unlinkSync(q.files.file.path);
         console.log("[Delete]-Successfully removed an empty file!");
+        s.send("<script>parent.UM.getEditor('"+editor_id+"').getWidgetCallback('image')('','ERROR')</script>");
+    }else if(!q.files.file){
+        s.send("<script>parent.UM.getEditor('"+editor_id+"').getWidgetCallback('image')('','ERROR')</script>");
     }else{
         target_path='/images/upload/'+(function(){return +new Date();})()+".jpg";
         fs.renameSync(q.files.file.path,"./public"+target_path);
